requirements.txt
streamlit==1.32.0
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.26
tavily-python>=0.3.0
trulens>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
